Moved AppSettings to Library (for ConnectionString access in Blazor Startup for EF Core) + update insertscript
