+++ IdeventAPI/Managers/ChipManager.cs
@@ -1,4 +1,5 @@
 ﻿using Dapper;
+using IdeventLibrary;
 using IdeventLibrary.Models;
 using System;
 using System.Collections.Generic;
