+++ IdeventAPI/Managers/EventStandManager.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Data;
 using System.Data.SqlClient;
 using Dapper;
+using IdeventLibrary;
 using IdeventLibrary.Models;
 
 namespace IdeventAPI.Managers
