+++ IdeventAPI/Managers/StandFunctionalityManager.cs
@@ -5,6 +5,7 @@ using System.Collections.Generic;
 using System.Data;
 using System.Data.SqlClient;
 using System.Linq;
+using IdeventLibrary;
 
 namespace IdeventAPI.Managers
 {
