+++ IdeventAPI/Managers/StandProductManager.cs
@@ -4,6 +4,7 @@ using IdeventLibrary.Models;
 using System.Collections.Generic;
 using System.Data;
 using System.Data.SqlClient;
+using IdeventLibrary;
 
 namespace IdeventAPI.Managers
 {
