+++ IdeventSQLServerTemplate/IdeventSQLServerTemplate.sqlproj
@@ -111,6 +111,7 @@
     <Build Include="dbo\Tables\AddressModel.sql" />
     <Build Include="dbo\StoredProcedures\spUpdateAddress.sql" />
     <Build Include="dbo\StoredProcedures\spUpdateCompany.sql" />
+    <Build Include="dbo\StoredProcedures\spGetAllProducts.sql" />
   </ItemGroup>
   <ItemGroup>
     <RefactorLog Include="IdeventSQLServerTemplate.refactorlog" />
